Call is Authorized function in the admin routes

+++ GalleryServer/GalleryServer/Areas/Admin/Features/Post/PostsController.cs
@@ -1,6 +1,7 @@
 ﻿namespace GalleryServer.Areas.Admin.Features.Post
 {
     using CloudinaryDotNet;
+    using GalleryServer.Areas.Admin.Features.Dashboard;
     using GalleryServer.Features.Cloudinary;
     using GalleryServer.Features.Post;
     using GalleryServer.Features.Post.Models;
@@ -14,29 +15,41 @@
         private readonly ICurrentUserService currentUser;
         private readonly ICloudinaryService cloudinaryService;
         private readonly Cloudinary cloudinary;
+        private readonly IAdminService adminService;
 
         public PostsController(            
             IPostsService posts,
             ICloudinaryService cloudinaryService,
             Cloudinary cloudinary,
-            ICurrentUserService currentUser)
+            ICurrentUserService currentUser,
+            IAdminService adminService)
         {            
             this.posts = posts;
             this.cloudinaryService = cloudinaryService;
             this.cloudinary = cloudinary;
             this.currentUser = currentUser;
+            this.adminService = adminService;
         }
 
-        public async Task<ActionResult> All()
+        [HttpGet("{page}")]
+        public async Task<ActionResult> All(int currentPage)
         {
-            var posts = this.posts.GetAllAdmin();
+            if (!await this.adminService.IsUserAdminAuthorized())
+            {
+                return BadRequest();
+            }
+            var posts = this.posts.GetAllAdmin(currentPage);
 
             return Accepted(nameof(this.All), posts);
         }        
 
         [HttpPost("{postId}")]
         public async Task<ActionResult> Update(string postId, [FromForm] UpdatePatchRequestModel model)
-        {       
+        {
+            if (!await this.adminService.IsUserAdminAuthorized())
+            {
+                return BadRequest();
+            }
             var userId = this.currentUser.GetId();            
             var pictureUrl = "";
 
@@ -58,6 +71,10 @@
         [HttpDelete("{postId}")]
         public async Task<ActionResult> Delete(string postId)
         {
+            if (!await this.adminService.IsUserAdminAuthorized())
+            {
+                return BadRequest();
+            }
             var userId = this.currentUser.GetId();
             var result = await this.posts.DeletePostAdmin(userId, postId);
